Move onActivityResult to proper activity. Remove redundant lambdas.
